[refactor] Api and file manager modules
Move all Schol-AR api and file manager related funcitonality into api_manager and sc_file_manager modules. General utilities for file actions used throughout the
project moved into utils module.
