+++ src/cmd.py
@@ -33,7 +33,9 @@ from typing import Union
 from chimerax.core.commands import CmdDesc, StringArg, BoolArg, SaveFileNameArg, SaveFolderNameArg, OpenFileNameArg
 from chimerax.core.commands import run
 
-from .io import ScARFileManager, APIManager
+from . import api_manager
+from . import sc_file_manager
+from .utils import empty_dir, get_first_file, path_exists, save_file_copy
 
 
 def login(session, username: str, api_token: Union[str, None] = None, **kwargs):
@@ -51,25 +53,25 @@ def login(session, username: str, api_token: Union[str, None] = None, **kwargs):
         session.logger.warning("Invalid username. Usernames can only contain letters, numbers, and spaces.")
         return
 
-    ScARFileManager.init_scholar_dirs()
+    sc_file_manager.init_scholar_dirs()
     user_token = api_token
     if user_token is None:
         # cmd call has implied that user already exists
-        retrieved_token = ScARFileManager.get_user_token(username)
+        retrieved_token = sc_file_manager.get_user_token(username)
         if retrieved_token is None:
             # if the user did not exist in the info file
             session.logger.warning(f"User {username} does not exist")
             return
         user_token = retrieved_token
 
-    if not APIManager.validate_api_token(user_token):
+    if not api_manager.validate_api_token(user_token):
         # exit on invalid api token
         session.logger.warning(f"Invalid API token for: {username}")
         return
 
-    ScARFileManager.update_users_info(username, user_token)
+    sc_file_manager.update_users_info(username, user_token)
 
-    ScARFileManager.update_user_projects(username)
+    sc_file_manager.update_user_projects(username)
     session.logger.status("Succesfully logged into Schol-AR as: " + username)
 
 
@@ -100,29 +102,29 @@ def project(session, username: str, project_title: str, project_type: str = "oth
         session.logger.warning("Invalid project title. Project titles can only contain letters, numbers, and spaces.")
         return
 
-    if project_type not in APIManager.PROJECT_TYPES.values():
-        project_types = ", ".join(APIManager.PROJECT_TYPES.values())
+    if project_type not in api_manager.PROJECT_TYPES.values():
+        project_types = ", ".join(api_manager.PROJECT_TYPES.values())
         session.logger.warning(f"Invalid project type. Project type must be one of: {project_types}")
         return
 
     # set target project if param project_title exists in the user's project list. None means it does not exist
-    target_project: dict = ScARFileManager.get_project(username, project_title)
+    target_project: dict = sc_file_manager.get_project(username, project_title)
 
     if target_project is None:
         # The project does not exist in the user's project list
         # create the project on server
-        token = ScARFileManager.get_user_token(username)
-        project_response = APIManager.create_project(token, project_title, project_type, disc_url)
+        token = sc_file_manager.get_user_token(username)
+        project_response = api_manager.create_project(token, project_title, project_type, disc_url)
         if project_response is None:
             # this will happen if there is a network call error
             session.logger.warning(f"Failed to create project: {project_title} on scholar server")
             return
 
         # create the project locally
-        ScARFileManager.update_user_projects(username)
+        sc_file_manager.update_user_projects(username)
 
         # now that the project is created try setting again
-        target_project = ScARFileManager.get_project(username, project_title)
+        target_project = sc_file_manager.get_project(username, project_title)
 
         # if the target_project is still none that means that there was an error creating the project locally
         if target_project is None:
@@ -130,7 +132,7 @@ def project(session, username: str, project_title: str, project_type: str = "oth
             return
 
     # Once we get to here we know that the project exists in the user's project list and qr dirs
-    ScARFileManager.update_augs_info(username, project_title)
+    sc_file_manager.update_augs_info(username, project_title)
 
 
 project_desc = CmdDesc(
@@ -153,23 +155,23 @@ def download_qr(session, username: str, project_title: str):
     if not usr_project_exists(username, project_title):
         return
 
-    qrstring = ScARFileManager.get_project_qrstring(username, project_title)
-    token = ScARFileManager.get_user_token(username)
-    qr_data = APIManager.get_qr_data(token, qrstring)
+    qrstring = sc_file_manager.get_project_qrstring(username, project_title)
+    token = sc_file_manager.get_user_token(username)
+    qr_data = api_manager.get_qr_data(token, qrstring)
 
     if qr_data is None:
         session.logger.warning(f"Failed to download project files for project {project_title}")
         return
 
-    pub_save_dir = ScARFileManager.pub_qr_dir(username, project_title)
-    admin_save_dir = ScARFileManager.admin_qr_dir(username, project_title)
+    pub_save_dir = sc_file_manager.pub_qr_dir(username, project_title)
+    admin_save_dir = sc_file_manager.admin_qr_dir(username, project_title)
 
     # if the qr code ever changes name for some reason or if there are extra files floating around this will clean up
-    ScARFileManager.empty_dir(pub_save_dir)
-    ScARFileManager.empty_dir(admin_save_dir)
+    empty_dir(pub_save_dir)
+    empty_dir(admin_save_dir)
 
-    APIManager.download_file_from_url(qr_data[APIManager.PUBLIC_QR_KEY], pub_save_dir)
-    APIManager.download_file_from_url(qr_data[APIManager.ADMIN_QR_KEY], admin_save_dir)
+    api_manager.download_file_from_url(qr_data[api_manager.PUBLIC_QR_KEY], pub_save_dir)
+    api_manager.download_file_from_url(qr_data[api_manager.ADMIN_QR_KEY], admin_save_dir)
 
 
 download_qr_desc = CmdDesc(
@@ -209,26 +211,26 @@ def augmentation(session, username: str, project_title: str, augmentation_title:
         return
 
     # check if the aug needs to be created or if it exists in the project info save file
-    target_aug_data = ScARFileManager.get_augmentation(username, project_title, augmentation_title)
+    target_aug_data = sc_file_manager.get_augmentation(username, project_title, augmentation_title)
     if target_aug_data is None:
         # The augmentation does not exist yet
 
-        token = ScARFileManager.get_user_token(username)
-        qrstring = ScARFileManager.get_project_qrstring(username, project_title)
+        token = sc_file_manager.get_user_token(username)
+        qrstring = sc_file_manager.get_project_qrstring(username, project_title)
 
         # check if the model and target image files are too large before creating the augmentation
-        if not ScARFileManager.save_and_size_check(session, '.glb'):
-            session.logger.warning(f"Model file must be smaller than {APIManager.MAX_FILE_SIZE_MB}MB. "
+        if not sc_file_manager.save_and_size_check(session, '.glb'):
+            session.logger.warning(f"Model file must be smaller than {api_manager.MAX_FILE_SIZE_MB}MB. "
                                    f"Reduce model detail and try again.")
             session.logger.warning("Could not create new augmentation.")
             return
-        if not ScARFileManager.save_and_size_check(session, '.png'):
-            session.logger.warning(f"Target Image file must be smaller than {APIManager.MAX_FILE_SIZE_MB}MB.")
+        if not sc_file_manager.save_and_size_check(session, '.png'):
+            session.logger.warning(f"Target Image file must be smaller than {api_manager.MAX_FILE_SIZE_MB}MB.")
             session.logger.warning("Could not create new augmentation.")
             return
 
         # The augmentation needs to be created on Schol-AR side.
-        create_aug_response = APIManager.create_augmentation(
+        create_aug_response = api_manager.create_augmentation(
             token, qrstring, augmentation_title, augmentation_type
         )
 
@@ -238,14 +240,14 @@ def augmentation(session, username: str, project_title: str, augmentation_title:
             return
 
         # update the project save file once we know the augmentation was created
-        ScARFileManager.update_augs_info(username, project_title)
+        sc_file_manager.update_augs_info(username, project_title)
 
         # refresh the target augmentation. Pulls from newly updated project info
-        target_aug_data = ScARFileManager.get_augmentation(username, project_title, augmentation_title)
+        target_aug_data = sc_file_manager.get_augmentation(username, project_title, augmentation_title)
 
         # This needs to happen before we try and upload the files to the server because we need to save them into the
         # directory structure
-        ScARFileManager.init_aug_dirs(username, project_title, augmentation_title, target_aug_data)
+        sc_file_manager.init_aug_dirs(username, project_title, augmentation_title, target_aug_data)
 
         run(session,
             f"scholar uploadAugFiles \"{username}\" \"{project_title}\" \"{augmentation_title}\" target_image True "
@@ -255,7 +257,7 @@ def augmentation(session, username: str, project_title: str, augmentation_title:
     else:
         # the augmentation already exists in the project info save file
         # make sure the dir structure for the augmentation is set up
-        ScARFileManager.init_aug_dirs(username, project_title, augmentation_title, target_aug_data)
+        sc_file_manager.init_aug_dirs(username, project_title, augmentation_title, target_aug_data)
 
 
 augmentation_desc = CmdDesc(
@@ -286,31 +288,31 @@ def download_aug_files(
         return
 
     if target_image:
-        target_image_url = ScARFileManager.get_augmentation_target_url(username, project_title, augmentation_title)
+        target_image_url = sc_file_manager.get_augmentation_target_url(username, project_title, augmentation_title)
 
         # check if the target image exists
         if target_image_url is None:
             session.logger.warning(f"Can't sync because Target Image for: {augmentation_title} not found")
         else:
             # only keep the 1 file that is being used
-            target_dir = ScARFileManager.aug_target_dir(username, project_title, augmentation_title)
-            ScARFileManager.empty_dir(target_dir)
+            target_dir = sc_file_manager.aug_target_dir(username, project_title, augmentation_title)
+            empty_dir(target_dir)
 
-            target_image_save_path = ScARFileManager.aug_target_dir(username, project_title, augmentation_title)
-            APIManager.download_file_from_url(target_image_url, target_image_save_path)
+            target_image_save_path = sc_file_manager.aug_target_dir(username, project_title, augmentation_title)
+            api_manager.download_file_from_url(target_image_url, target_image_save_path)
 
     if augmented_file:
-        augmented_url = ScARFileManager.get_augmentation_model_url(username, project_title, augmentation_title)
+        augmented_url = sc_file_manager.get_augmentation_model_url(username, project_title, augmentation_title)
         # check if the augmented file exists
         if augmented_url is None:
             session.logger.warning(f"Can't sync because Augmented File for: {augmentation_title} not found")
         else:
             # only keep the 1 file that is being used in the directory
-            model_dir = ScARFileManager.aug_model_dir(username, project_title, augmentation_title)
-            ScARFileManager.empty_dir(model_dir)
+            model_dir = sc_file_manager.aug_model_dir(username, project_title, augmentation_title)
+            empty_dir(model_dir)
 
-            augmented_save_path = ScARFileManager.aug_model_dir(username, project_title, augmentation_title)
-            APIManager.download_file_from_url(augmented_url, augmented_save_path)
+            augmented_save_path = sc_file_manager.aug_model_dir(username, project_title, augmentation_title)
+            api_manager.download_file_from_url(augmented_url, augmented_save_path)
 
 
 download_aug_files_desc = CmdDesc(
@@ -341,20 +343,20 @@ def upload_aug_files(session, username: str, project_title: str, augmentation_ti
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
 
-    token = ScARFileManager.get_user_token(username)
+    token = sc_file_manager.get_user_token(username)
 
     # it is important that augmented file gets patched before the target image does. As of 07/30/24 schol-ar has a bug
     # where if the target image is updated directly before the model it will get stuck displayed as processing.
     if augmented_file:
-        model_dir = ScARFileManager.aug_model_dir(username, project_title, augmentation_title)
-        ScARFileManager.empty_dir(model_dir)
-        model_file = ScARFileManager.aug_model_file(username, project_title, augmentation_title)
+        model_dir = sc_file_manager.aug_model_dir(username, project_title, augmentation_title)
+        empty_dir(model_dir)
+        model_file = sc_file_manager.aug_model_file(username, project_title, augmentation_title)
         aug_save_and_update(session, token, username, project_title, augmentation_title, model_file,
                             target_update=False, verbose=verbose)
     if target_image:
-        target_image_dir = ScARFileManager.aug_target_dir(username, project_title, augmentation_title)
-        ScARFileManager.empty_dir(target_image_dir)
-        target_image_file = ScARFileManager.aug_target_file(username, project_title, augmentation_title)
+        target_image_dir = sc_file_manager.aug_target_dir(username, project_title, augmentation_title)
+        empty_dir(target_image_dir)
+        target_image_file = sc_file_manager.aug_target_file(username, project_title, augmentation_title)
         aug_save_and_update(session, token, username, project_title, augmentation_title, target_image_file,
                             target_update=True, verbose=verbose)
 
@@ -384,12 +386,12 @@ def aug_save_and_update(session, token: str, username: str, project_title: str, 
     :param target_update: If True, update the target image. If False, update the augmented file.
     :param verbose: If True, the command will log additional information. Default is False.
     """
-    qrstring = ScARFileManager.get_project_qrstring(username, project_title)
-    aug_id = ScARFileManager.get_augmentation_id(username, project_title, augmentation_title)
+    qrstring = sc_file_manager.get_project_qrstring(username, project_title)
+    aug_id = sc_file_manager.get_augmentation_id(username, project_title, augmentation_title)
     # save the file to local directory
     run(session, f"save \"{file_path}\"", log=verbose)
     # update the augmentation on the server
-    updated_aug_info = APIManager.edit_augmentation(
+    updated_aug_info = api_manager.edit_augmentation(
         token, qrstring, aug_id, file_path, target_update=target_update
     )
     # something on the server failed check
@@ -399,7 +401,7 @@ def aug_save_and_update(session, token: str, username: str, project_title: str, 
                                f"{project_title}")
         return
     # make sure that the save files for the project are updated
-    ScARFileManager.update_augs_info(username, project_title)
+    sc_file_manager.update_augs_info(username, project_title)
 
 
 def save_aug_session(session, username: str, project_title: str, augmentation_title: str,
@@ -418,18 +420,18 @@ def save_aug_session(session, username: str, project_title: str, augmentation_ti
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
 
-    aug_session_dir = ScARFileManager.aug_session_dir(username, project_title, augmentation_title)
+    aug_session_dir = sc_file_manager.aug_session_dir(username, project_title, augmentation_title)
 
     # if the file path is not valid, save the existing session to the augmentation and set the target session path
-    if file_path is None or not ScARFileManager.path_exists(file_path):
-        target_session_file = ScARFileManager.aug_session_file(username, project_title, augmentation_title)
+    if file_path is None or not path_exists(file_path):
+        target_session_file = sc_file_manager.aug_session_file(username, project_title, augmentation_title)
         # empty the directory before saving the session
-        ScARFileManager.empty_dir(aug_session_dir)
+        empty_dir(aug_session_dir)
         run(session, f"save \"{target_session_file}\"", log=verbose)
     else:
         # Valid session file given from outside Schol-AR dir structure. Copy the file to the augmentation session dir.
-        ScARFileManager.empty_dir(aug_session_dir)
-        ScARFileManager.save_file_copy(file_path, aug_session_dir)
+        empty_dir(aug_session_dir)
+        save_file_copy(file_path, aug_session_dir)
 
 
 save_aug_session_desc = CmdDesc(
@@ -457,8 +459,8 @@ def open_aug_session(session, username: str, project_title: str, augmentation_ti
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
 
-    aug_session_dir = ScARFileManager.aug_session_dir(username, project_title, augmentation_title)
-    session_file_name = ScARFileManager.get_first_file(aug_session_dir)
+    aug_session_dir = sc_file_manager.aug_session_dir(username, project_title, augmentation_title)
+    session_file_name = get_first_file(aug_session_dir)
     if session_file_name is None:
         session.logger.info(f"No session file yet for Augmentation: {augmentation_title}")
         return
@@ -491,12 +493,12 @@ def store_target_image(session, username: str, project_title: str, augmentation_
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
 
-    target_image_path = ScARFileManager.get_augmentation_target_image_path(username, project_title, augmentation_title)
+    target_image_path = sc_file_manager.get_augmentation_target_image_path(username, project_title, augmentation_title)
     if target_image_path is None:
         run(session,
             f"scholar downloadAugFiles \"{username}\" \"{project_title}\" \"{augmentation_title}\" "
             f"targetImage True augmentedFile False", log=verbose)
-        target_image_path = ScARFileManager.get_augmentation_target_image_path(
+        target_image_path = sc_file_manager.get_augmentation_target_image_path(
             username, project_title, augmentation_title)
 
     if target_image_path is not None:
@@ -529,11 +531,11 @@ def store_model(session, username: str, project_title: str, augmentation_title: 
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
 
-    model_path = ScARFileManager.get_auggmentation_model_file_path(username, project_title, augmentation_title)
+    model_path = sc_file_manager.get_auggmentation_model_file_path(username, project_title, augmentation_title)
     if model_path is None:
         run(session, f"scholar downloadAugFiles \"{username}\" \"{project_title}\" \"{augmentation_title}\" "
                      f"targetImage False augmentedFile True", log=verbose)
-        model_path = ScARFileManager.get_auggmentation_model_file_path(username, project_title, augmentation_title)
+        model_path = sc_file_manager.get_auggmentation_model_file_path(username, project_title, augmentation_title)
 
     if model_path is not None:
         # if the model path exists, and the save directory exists, copy the file
@@ -566,7 +568,7 @@ def store_all_aug_files(session, username: str, project_title: str, augmentation
     if not usr_proj_aug_exists(username, project_title, augmentation_title):
         return
     os.makedirs(save_folder, exist_ok=True)
-    safe_aug_filename = APIManager.sanitize_file_name(augmentation_title)
+    safe_aug_filename = api_manager.sanitize_file_name(augmentation_title)
     model_file_path = os.path.join(save_folder, f"{safe_aug_filename}.glb")
     target_image_file_path = os.path.join(save_folder, f"{safe_aug_filename}.png")
     # Use the project title as the qr image file name. The qr title is a unique identifier and is confusing.
@@ -601,10 +603,10 @@ def store_qr_image(session, username: str, project_title: str, save_location: st
     if not usr_project_exists(username, project_title):
         return
 
-    pub_qr_image_path = ScARFileManager.get_qr_file(username, project_title, admin=False)
+    pub_qr_image_path = sc_file_manager.get_qr_file(username, project_title, admin=False)
     if pub_qr_image_path is None:
         run(session, f"scholar downloadQR \"{username}\" \"{project_title}\"", log=verbose)
-        pub_qr_image_path = ScARFileManager.get_qr_file(username, project_title, admin=False)
+        pub_qr_image_path = sc_file_manager.get_qr_file(username, project_title, admin=False)
 
     if pub_qr_image_path is not None:
         # if the public QR code image path exists, and the save directory exists, copy the file
@@ -645,13 +647,13 @@ def clean_local(session, username: str = None):
     target_usernames = [username]
 
     if username is None:
-        target_usernames = ScARFileManager.list_usernames()
+        target_usernames = sc_file_manager.list_usernames()
     elif not username_exists(username):
         session.logger.warning(f"Cannot Clean Local. User {username} not found")
         return
 
     for username in target_usernames:
-        ScARFileManager.clean_local(username)
+        sc_file_manager.clean_local(username)
 
 
 clean_local_desc = CmdDesc(
@@ -669,7 +671,7 @@ def remove_user(session, username: str):
     :param username: The username of the Schol-AR user.
     """
 
-    if ScARFileManager.remove_user(username):
+    if sc_file_manager.remove_user(username):
         session.logger.info(f"User {username} removed")
         return
     session.logger.warning(f"Can't remove user {username} because it was not found")
@@ -689,7 +691,7 @@ def username_exists(username: str):
     :return: True if the username exists, False otherwise.
     """
     # make sure the user exists
-    if not ScARFileManager.username_exists(username):
+    if not sc_file_manager.username_exists(username):
         print(f"User {username} not found")
         return False
     return True
@@ -706,7 +708,7 @@ def usr_project_exists(username: str, project_title: str):
     # make sure the user and project exist
     if not username_exists(username):
         return False
-    if not ScARFileManager.project_exists(username, project_title):
+    if not sc_file_manager.project_exists(username, project_title):
         print(f"Project {project_title} not found")
         return False
     return True
@@ -724,7 +726,7 @@ def usr_proj_aug_exists(username: str, project_title: str, augmentation_title: s
     # make sure the user, project, and augmentation exist
     if not usr_project_exists(username, project_title):
         return False
-    if not ScARFileManager.aug_exists(username, project_title, augmentation_title):
+    if not sc_file_manager.aug_exists(username, project_title, augmentation_title):
         print(f"Augmentation {augmentation_title} not found")
         return False
     return True
